autobot/requirements.txt
altgraph==0.17.3
et-xmlfile==1.1.0
future==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.7
jupyter_core==5.1.0
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
papermill==2.4.0
pefile==2022.5.30
Pillow==9.2.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyparsing==3.0.9
PyPDF4==1.27.0
python-dateutil==2.8.2
python-minifier==2.6.0
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
reportlab==3.6.11
rpa==1.48.1
six==1.16.0
tagui==1.48.1
xlwings==0.28.5